# [FIX] base: improve performance on res.device model