Issue:
-------

The `res.device.log` model contains a huge number of records and this causes a performance problem
for users belonging to the `base.group_system` group.

The `res.device` table is a non-materialized view defined with a `DISTINCT ON` clause
that uses the `user_id`, `session_identifier`, `platform`, `browser` columns.

When we want to retrieve the `res.device` records for a user, two queries are performed.
The first will retrieve the `res.device` ids for the user.
```sql
SELECT d.id FROM res_device d WHERE d.user_id=<user_id>
```
The second will retrieve the information for the `res.device` found in the first query.
```sql
SELECT d.* FROM res_device WHERE d.id IN (<device_ids>)
```

The second query will be very slow for an administrator, because Postgresql cannot use
the index on the primary key (`id`). Only the fields used in the `DISTINCT ON` clause
can be used as an index.
As a result, Postgresql will have to scan the entire `res_device_log` table, which is very large.

For a user who does not belong to the `base.group_system` group, an `ir.rule` is applied
so that they can only see their own `res_device`. For them, the second query is:
```sql
SELECT d.* FROM res_device WHERE d.id IN (<device_ids>) AND d.user_id = <user_id>
```
Postgresql can use the index on `user_id` to filter the table very quickly.
Then it is very fast to find the devices corresponding to the ids in the reduced table.

Solution:
----------

Unfortunately, it doesn't make sense to add the `id` column to the `DISTINCT ON` clause.
One solution could be to keep the `ir.rule` for people in the `base.group_system` group,
but we want administrators to have access to the `res.device` records of all users.
This solution was therefore rejected.

The solution adopted is to modify the non-materialized view so that Postgresql can use
the index on the primary key.

The goal of the view is to deduplicate records from the `res.device.log` model.
If for a given record, we find another record which has the same `user_id`, `session_identifier`,
`platform`, `browser` with a more recent `last_activity`, then we don't keep it.
Add a comparison between the `id` if the `last_activity` values are the same.

In addition, we need to take into account the case where we are unable to determine
the `platform` and the `browser` with the user agent.
This leads to `NULL` values, which must be handled with `IS DISTINCT FROM`.

To improve performance, we can also add a composite index with `user_id`, `session_identifier`,
`platform` and `browser` on `res.device.log`.

Appendices:
------------

Non administrator:
```
EXPLAIN ANALYZE SELECT * FROM res_device WHERE id=1 AND user_id=2;
 Subquery Scan on res_device  (cost=4.46..4.48 rows=1 width=126) (actual time=0.114..0.125 rows=0 loops=1)
   Filter: (res_device.id = 1)
   Rows Removed by Filter: 2
   ->  Unique  (cost=4.46..4.47 rows=1 width=126) (actual time=0.089..0.113 rows=2 loops=1)
         ->  Sort  (cost=4.46..4.46 rows=1 width=126) (actual time=0.086..0.095 rows=3 loops=1)
               Sort Key: d.session_identifier, d.platform, d.browser, d.last_activity DESC, d.id DESC
               Sort Method: quicksort  Memory: 25kB
               ->  Index Scan using res_device_log__user_id_index on res_device_log d  (cost=0.43..4.45 rows=1 width=126) (actual time=0.049..0.064 rows=3 loops=1)
                     Index Cond: (user_id = 2)
                     Filter: (NOT revoked)
 Planning Time: 0.224 ms
 Execution Time: 0.171 ms
```

Administrator:
```
EXPLAIN ANALYZE SELECT * FROM res_device WHERE id=1;
 Subquery Scan on res_device  (cost=475345.64..1031211.97 rows=20000 width=126) (actual time=32385.982..32425.079 rows=0 loops=1)
   Filter: (res_device.id = 1)
   Rows Removed by Filter: 3999995
   ->  Unique  (cost=475345.64..981211.93 rows=4000003 width=126) (actual time=5740.682..27526.236 rows=3999995 loops=1)
         ->  Gather Merge  (cost=475345.64..941211.90 rows=4000003 width=126) (actual time=5740.679..17382.463 rows=4000003 loops=1)
               Workers Planned: 2
               Workers Launched: 2
               ->  Sort  (cost=474345.61..478512.28 rows=1666668 width=126) (actual time=5723.180..8743.966 rows=1333334 loops=3)
                     Sort Key: d.user_id, d.session_identifier, d.platform, d.browser, d.last_activity DESC, d.id DESC
                     Sort Method: external merge  Disk: 159344kB
                     Worker 0:  Sort Method: external merge  Disk: 158864kB
                     Worker 1:  Sort Method: external merge  Disk: 159360kB
                     ->  Parallel Seq Scan on res_device_log d  (cost=0.00..85632.68 rows=1666668 width=126) (actual time=74.375..1997.856 rows=1333334 loops=3)
                           Filter: (NOT revoked)
 Planning Time: 0.215 ms
 JIT:
   Functions: 9
   Options: Inlining true, Optimization true, Expressions true, Deforming true
   Timing: Generation 1.220 ms, Inlining 84.317 ms, Optimization 86.161 ms, Emission 52.570 ms, Total 224.268 ms
 Execution Time: 32443.730 ms
```

Administrator with corrected view:
```
EXPLAIN ANALYZE SELECT * FROM res_device WHERE id=1;
 Sort  (cost=16.93..16.94 rows=1 width=126) (actual time=0.854..0.864 rows=0 loops=1)
   Sort Key: d.last_activity DESC
   Sort Method: quicksort  Memory: 25kB
   ->  Nested Loop Anti Join  (cost=0.86..16.92 rows=1 width=126) (actual time=0.844..0.852 rows=0 loops=1)
         ->  Index Scan using res_device_log_pkey on res_device_log d  (cost=0.43..8.45 rows=1 width=126) (actual time=0.608..0.613 rows=1 loops=1)
               Index Cond: (id = 1)
               Filter: (NOT revoked)
         ->  Index Scan using res_device_log__session_identifier_index on res_device_log d2  (cost=0.43..8.46 rows=1 width=49) (actual time=0.221..0.223 rows=1 loops=1)
               Index Cond: ((session_identifier)::text = (d.session_identifier)::text)
               Filter: ((NOT revoked) AND (NOT ((platform)::text IS DISTINCT FROM (d.platform)::text)) AND (NOT ((browser)::text IS DISTINCT FROM (d.browser)::text)) AND (user_id = d.user_id) AND ((last_activity > d.last_activity) OR ((last_activity = d.last_activity) AND (id > d.id))))
               Rows Removed by Filter: 1
 Planning Time: 8.539 ms
 Execution Time: 0.930 ms
```

Administrator with corrected view and index:
```
EXPLAIN ANALYZE SELECT * FROM res_device WHERE id=1;
 Sort  (cost=16.93..16.94 rows=1 width=126) (actual time=0.151..0.163 rows=0 loops=1)
   Sort Key: d.last_activity DESC
   Sort Method: quicksort  Memory: 25kB
   ->  Nested Loop Anti Join  (cost=0.86..16.92 rows=1 width=126) (actual time=0.140..0.149 rows=0 loops=1)
         ->  Index Scan using res_device_log_pkey on res_device_log d  (cost=0.43..8.45 rows=1 width=126) (actual time=0.063..0.067 rows=1 loops=1)
               Index Cond: (id = 1)
               Filter: (NOT revoked)
         ->  Index Only Scan using res_device_log__composite_idx on res_device_log d2  (cost=0.43..8.46 rows=1 width=49) (actual time=0.064..0.066 rows=1 loops=1)
               Index Cond: ((user_id = d.user_id) AND (session_identifier = (d.session_identifier)::text))
               Filter: ((NOT ((platform)::text IS DISTINCT FROM (d.platform)::text)) AND (NOT ((browser)::text IS DISTINCT FROM (d.browser)::text)) AND ((last_activity > d.last_activity) OR ((last_activity = d.last_activity) AND (id > d.id))))
               Rows Removed by Filter: 1
               Heap Fetches: 1
 Planning Time: 11.480 ms
 Execution Time: 0.225 ms
```