@robodoo 2024-11-04
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/186106.png)](https://mergebot.odoo.com/odoo/odoo/pull/186106)